move tests from bpf_ep_mgr_connlimit_test.go to bpf_ep_mgr_test.go
delete bpf_ep_mgr_connlimit_test.go
